Finished implementing SignInWithGoogle AB#31
