+++ src/SeeSharp.Api/appsettings.json
@@ -1,46 +1,46 @@
 ﻿{
-    "ConnectionStrings": {
-        "DefaultConnection": "Server=localhost,1433;Database=SeeSharpBlog;User ID=sa;Password=P@ssw0rd;TrustServerCertificate=True"
-        // "DefaultConnection": "Server=.\\SQLExpress;Database=SeeSharpBlog;Trusted_Connection=True;TrustServerCertificate=True"
+  "ConnectionStrings": {
+    //"DefaultConnection": "Server=localhost,1433;Database=SeeSharpBlog;User ID=sa;Password=P@ssw0rd;TrustServerCertificate=True"
+    "DefaultConnection": "Server=.\\SQLExpress;Database=SeeSharpBlog;Trusted_Connection=True;TrustServerCertificate=True"
+  },
+  "Serilog": {
+    "Using": [ "Serilog.Sinks.Console", "Serilog.Sinks.File" ],
+    "MinimumLevel": {
+      "Default": "Information",
+      "Override": {
+        "Microsoft": "Warning",
+        "System": "Warning"
+      }
     },
-    "Serilog": {
-        "Using": [ "Serilog.Sinks.Console", "Serilog.Sinks.File" ],
-        "MinimumLevel": {
-            "Default": "Information",
-            "Override": {
-                "Microsoft": "Warning",
-                "System": "Warning"
-            }
-        },
-        "WriteTo": [
-            { "Name": "Console" },
-            {
-                "Name": "File",
-                "Args": {
-                    "path": "/logs/log-.txt",
-                    "rollingInterval": "Day",
-                    "rollOnFileSizeLimit": true,
-                    "formatter": "Serilog.Formatting.Compact.CompactJsonFormatter"
-                }
-            }
-        ],
-        "Enrich": [ "FromLogContext", "WithMachineName", "WithThreadId" ]
-    },
-    "Logging": {
-        "LogLevel": {
-            "Default": "Information",
-            "Microsoft.AspNetCore": "Warning"
+    "WriteTo": [
+      { "Name": "Console" },
+      {
+        "Name": "File",
+        "Args": {
+          "path": "/logs/log-.txt",
+          "rollingInterval": "Day",
+          "rollOnFileSizeLimit": true,
+          "formatter": "Serilog.Formatting.Compact.CompactJsonFormatter"
         }
-    },
-    "JwtOptions": {
-        "Secret": "C#-SeeSharp-JwtKey",
-        "Issuer": "C#-SeeSharp",
-        "Audience": "https://localhost:4200",
-        "expiryInMinutes": 60
-    },
-    "GoogleAuthOptions": {
-        "ClientId": "[From https://console.cloud.google.com/]",
-        "ClientSecret": "[From https://console.cloud.google.com/]"
-    },
-    "AllowedHosts": "*"
+      }
+    ],
+    "Enrich": [ "FromLogContext", "WithMachineName", "WithThreadId" ]
+  },
+  "Logging": {
+    "LogLevel": {
+      "Default": "Information",
+      "Microsoft.AspNetCore": "Warning"
+    }
+  },
+  "JwtOptions": {
+    "Secret": "C#-SeeSharp-JwtKey",
+    "Issuer": "C#-SeeSharp",
+    "Audience": "https://localhost:4200",
+    "expiryInMinutes": 60
+  },
+  "GoogleAuthOptions": {
+    "ClientId": "[From https://console.cloud.google.com/]",
+    "ClientSecret": "[From https://console.cloud.google.com/]"
+  },
+  "AllowedHosts": "*"
 }
