+++ src/SeeSharp.Application/Common/Interfaces/IIdentityService.cs
@@ -10,6 +10,8 @@ public interface IIdentityService
     Task<(string UserId,string FullName, string UserName, string Email, IList<string> Roles)> GetUserDetailsByUserNameAsync(string userName);
 
     Task<(string UserId, string FullName, string UserName, string Email, IList<string> Roles)> GetUserDetailsByEmailAsync(string userName);
+    
+    Task<bool> UserExists(string email);
 
     Task<Result> AddToRolesAsync(string userId, List<string> roles);
