+++ src/SeeSharp.Application/Features/Authentication/Queries/ValidateJwtQuery.cs
