+++ src/SeeSharp.Domain/Constants/Roles.cs
@@ -8,4 +8,6 @@ namespace SeeSharp.Domain.Constants;
 public abstract class Roles
 {
     public const string Administrator = nameof(Administrator);
+    public const string Contributor = nameof(Contributor);
+    public const string BasicUser = nameof(BasicUser);
 }
